:pencil: se divide codigo para mejor manejo, para rutas se llama a la función en consultas destinada para la ruta get /joyas
